+++ player.hpp
@@ -6,15 +6,15 @@
 class Player {
     public:
         static const int SIZE;
-        static const float DEFAULT_DX;
-        static const float DEFAULT_DY;
+        static const double DEFAULT_DX;
+        static const double DEFAULT_DY;
     protected:
         int left;
         int top;
         ALLEGRO_BITMAP *bitmap;
         int score;
-        float dx;
-        float dy;
+        double dx;
+        double dy;
     
     public:
         Player(int top, int left, ALLEGRO_BITMAP *bitmap);
@@ -23,8 +23,8 @@ class Player {
         int getRight();
         int getBottom();
         int getScore();
-        float getDx();
-        float getDy();
+        double getDx();
+        double getDy();
         void setLeft(int left);
         void setTop(int top);
         void incScore();
